画像のアップロードとTipsの投稿を1つのボタンで同時に行う
