+++ hacku-team1/src/components/molecules/RecommendDialog.tsx
@@ -8,8 +8,9 @@ import {
   DialogTitle,
   TextField,
 } from "@material-ui/core";
-import Image from "../atoms/Image";
-import { db } from "../../firebase/firebase";
+
+import firebase from "firebase/app";
+import { db, firebaseApp } from "../../firebase/firebase";
 import Upload from "../molecules/Upload";
 
 type Props = {
@@ -17,6 +18,11 @@ type Props = {
   title: string;
   isOpen: boolean;
   doClose: () => void;
+};
+
+// previewを追加
+type MyFile = File & {
+  preview: string;
 };
 
 const useStyles = makeStyles((theme) => ({
@@ -33,27 +39,100 @@ const RecommendDialog = (props: Props) => {
 
   const [open, setOpen] = useState(false);
   const [input, setInput] = useState("");
-  const [imageURL, setImageURL] = useState("");
+  const [imageURL, setImageURL] = useState<any>(""); // Promissの型の指定が複雑なので後回し
+  const [uploading, setUploading] = useState(false);
+  const [progress, setProgress] = useState(0);
+  const [files, setFiles] = useState<MyFile[]>([]);
 
   useEffect(() => {
     setOpen(props.isOpen);
   }, [props.isOpen]);
 
   const uploadData = async () => {
-    const userRef = db.collection("Tips").doc(props.docid);
-    await userRef.update({
-      done: true,
-      doneContent: input,
-      imageURL: imageURL,
+
+    console.log("Image onUpload start");
+
+    // ローディングをOn。progressを初期化
+    setUploading(true);
+    setProgress(0);
+
+    function uploadImageAsPromise(file: File) {
+      console.log("uploadImageAsPromise start");
+
+      // アップロード先のファイルパスの作成
+      const file_name = file.name;
+      const storageRef = firebaseApp
+        .storage()
+        .ref()
+        .child("images/" + file_name);
+
+      return new Promise(function (resolve, reject) {
+        //Upload file
+        var task = storageRef.put(file);
+
+        //Update progress bar
+        task.on(
+          firebase.storage.TaskEvent.STATE_CHANGED,
+          function progress(snapshot) {
+            var percent =
+              (snapshot.bytesTransferred / snapshot.totalBytes) * 100;
+            console.log(percent + "% done");
+          },
+          function error(err) {
+            // 失敗時
+            console.log("upload error");
+            reject(err);
+          },
+          function complete() {
+            // 成功時
+            console.log("upload complete.");
+            task.then(function (snapshot: firebase.storage.UploadTaskSnapshot) {
+              resolve(snapshot.ref.getDownloadURL());
+            });
+          }
+        );
+      })
+        .then(function (downloadURL) {
+          console.log("Finished uploading file: " + file_name);
+
+          // progressを更新する
+          setProgress((oldProgress) => oldProgress + 1);
+          return downloadURL;
+        })
+        .catch(function () {
+          console.log("Error:uploadImageAsPromise");
+        });
+    }
+
+    // 複数のファイルアップロードをPromise.allで並列に実行する
+    const file = files[0]
+    // const result = (new Promise(uploadImageAsPromise(file)));
+    const result = uploadImageAsPromise(file)
+    result.then( async function (imageURL) {
+      console.log("Upload result");
+      console.log(imageURL);
+      setImageURL(imageURL);
+
+      const userRef = db.collection("Tips").doc(props.docid);
+      await userRef.update({
+        done: true,
+        doneContent: input,
+        imageURL: imageURL,
+      }).then(() => {
+        // 最後にいろいろな変数のリセット
+        setUploading(false);
+        setProgress(0);
+        setFiles([]);
+        setInput("");
+        window.location.reload();
+        setOpen(false);
+        props.doClose();
+      })
     });
   };
 
   const handleCloseWithUpload = () => {
     uploadData();
-    window.location.reload();
-    setInput("");
-    setOpen(false);
-    props.doClose();
   };
 
   const handleCloseWithCancel = () => {
@@ -74,7 +153,12 @@ const RecommendDialog = (props: Props) => {
         </DialogTitle>
         <DialogContent>
           <div className={classes.content}>
-            <Upload setImageURL={setImageURL} />
+            <Upload 
+              uploading={uploading}
+              files={files}
+              progress={progress}
+              setImageURL={setImageURL}
+              setFiles={setFiles} />
           </div>
           <form>
             <TextField
