+++ hacku-team1/src/components/pages/TopPage.tsx
@@ -79,7 +79,6 @@ const TopPage: React.FC<Props> = ({ uid }) => {
         });
       });
     setContentDataList(temporaryContentData);
-    console.log(await Promise.all(contentDataList));
   };
 
   return (
